requirements.txt
Django>=3.2,<5.0
instaloader>=4.9
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
pydrive>=1.3.1
python-dotenv>=1.0.0
Pillow>=9.0.0
httpx>=0.23.0
mysqlclient==2.1.0
selenium==4.19.0
webdriver-manager==4.0.1
